Ensured email is not required to authenticate users
The OpenID Connect defines the 'email' scope as an optional scope. So the 'email' value is not necessarily present in the user info data returned by the OpenID Connect
provider ; this should prevent users from being authenticated if the client did not ask for the 'email' scope during the flow.